Make all PackageWidget subclasses use the proxymodel filtering facilities rather than using setPackages with a QApt::PackageList. By setting the filter before setting the packages to QApt::Backend::availablePackages() we prevent performance losses during the population of ReviewWidget and UpdaterWidget.
This makes everything safer and quashes the "marking $package for $action crashes muon" crashes.

A xapian indexing progress bar for the status bar also snuck in to this commit, because I was still working on it and wasn't quite ready to commit. But now the changes are intermingled so I'm committing it. It still has a few issues...

svn path=/trunk/extragear/sysadmin/muon/; revision=1178693