Switch to a white background for the ApplicationDetailsWidget.
This solves several problems; I need to use a lighter font color for some labels
in the Review widgets, and the separator I was using to differentiate the nav/search
bar didn't look good if somebody uses the Oxygen window decoration's separator.

It would be neat to get a subtle background image, like the Ubuntu Software Center uses.