Exit synchronously in the case of a backend initialization error.
The way that the asynchronous error reporting currently works will not close the
application "quick" enough when QApplication::exit is called. In the future, the smart
thing to do would to make QApt::Backend::reloadCache return a bool, but that would
break ABI and will have to wait for QApt2.

BUG:287988
FIXED-IN:1.3.0