Provide a new ratings metric for the Application
It's called RatingPoints and it's based on the histogram of reviews:
1 star reviews get 1 point, 2 star get 2 points, etc...

This way we get to value also the applications that have more ratings too,
which is is interesting because it shows where users focus is.

The algorithm choosed to make this wasn't really thought through, maybe we
can find a better one, it's good for the moment, though. I think.

CCMAIL: eskerda@gmail.com