Revert "The Package::controlField() checks we were doing in various functions for extras.ubuntu.com packages are kinda expensive."
This reverts commit d67754543d8c55ab673159e45a4be10019cece98.

Conflicts:

	libmuon/Application.cpp

I reverted this commit because it makes the ApplicationModelTest
not to pass by crashing. This means that for some reason it puts application
instances in a weird state.

I'm unsure of what you were trying to attempt and I'd prefer you to review the change
and fix either the test or the new code.

Please review it and the next time check if the tests pass before committing.

CCMAIL: echidnaman@kubuntu.org