Set a configuration KConfig file to provide the app some data.
The idea is that the packager should be able to provide some data
to the application to specify some distro specific sources.

At the moment there are 3 of them: it lets you choose what debian
screenshots instance you'd like to pull your information from, also
the ratings and reviews server and the featured apps.