Sort the items within the update categories. (Using a nifty c++11 lambda to pass to qSort!)
This also had the side-effect of exposing a bug in the "forensic" implementation of UpdateItem::type()
Instead of figuring out our own type due to parentage (or lack thereof), set the type in the constructor since the types have unique constructors, and return that.

As a bonus, the ItemType enum is now a c++11 enum class. This makes the enum items harder to misuse, as well as allowing me to define the underlying enum type as 8-bit rather than an int for some extra memory savings. (A triple-hit feature, bugfix, and optimization patch!)