Exceptions that MetaReleaseCore throw during a download happen in a separate thread, so we can't catch them.
The root cause of the IOError seems to be the existence of a proxy, and it appears that UpdateManager has a proxy init function. (Poor design on UpdateManager's part...)
This should fix the root cause of this bug, but I fear for other exceptions that could be thrown that MetaReleaseCore won't catch for us...
I'll commit a separate fix to make our KProcess call to releasechecker non-blocking.

CCBUG:307049