Refactor QApt::Backend initialization in ApplicationBackend
Merged initBackend and setBackend. initBackend was the only user of
setBackend and having the API could have meant that it can be set, make
it explicit.

If the backend hasn't been initialized, don't initialize QAptActions yet,
wait until it has been initialized.
This fixes a crash when loading the muon-appsbackend with gui integration.