Return QUrl() if there's no screenshot to offer
If the application is told that there's no screenshot,
return an empty screenshot. This will help make the UI more
responsible by reacting to the lack of screenshot instead
of waiting for it to be fetched. It also improves the
loading time of an applications grid, since we get to
reduce some http calls.