Move the LaunchListModel to muon installer
This way we can manage from there all the launch specific states. Now
the model is instanciated once at the beginning of the run and it will
be responsible for tracking all executables installations.
This commit also removes some API that wasn't being really used inside
the ApplicationBackend that was responsible for keeping track of the
launches list. This won't happen in the backend anymore.