Cache root categories
Don't re-parse the categories every time we set a CategoriesModel to root.
This makes it possible to easily set up different CategoriesModels without
having a throughput constraint.

PS: It would be interesting to reconsider the role of the CategoryModel, to
me it seems like we're doing both the model and the proxy at the same time.