Improve the way we start discover from a category
Move the category lookup to the CategoryModel, extract it from discover's
qml code. It simplifies the implementation and makes it more flexible at
the same time.

This is important because we were instantiating a broken Category instance,
and now we're using Categories as expected, by finding it within the
populated categories from the xml file.