Use storedGet instead of putting the data in a temporary file
It reduces the complexity of the code a bit and uses better resources
by not putting something on the disk for reading it right away.
I left the jobs in a set so that they all can be cancelled at the
same time, but probably it would be a good idea to get rid of that too.