Move the updater APT-dependent actions initialization to the backend
To do so by keeping it the same way, the updaters will return a list of
actions. The actions with higher priority will be put as KMessageWidgets
and the rest as tool bar buttons.
The StandardUpdater won't report any actions and the actions have been
moved from updater to ApplicationUpdates (those will have to be tested
later on).