Improve and update how ApplicationsTop sizes are calculated
Since some recent changes in plasma themes the Label stopped looking
correct, fix it by setting the size to the header label ourselves.
Also add a small margin between the icon and the resource name, because it
didn't look very well.
Calculate the height of the items instead of doing weird guessing