Fix a crash on exit, along with an additional crash on exit that fixing the first crash exposed. This was done by:  - Not making ResourcesModel a global static via K_GLOBAL_STATIC in order to make it singleton. ResourcesModel was being kept around after the QApplication's destruction, which meant that some of the children within ResourcesModel's QObject tree had already been deleted. This caused a "pure virtual method called" when ResourcesModel was destructed, especially if another thread had been using the QApt::Backend at the time when the QApplication was destroyed.  - Only allowing the KApplication to close until after its initialization thread had finished.
For the first bug, the ResourcesModel was being kept around after the QApplication's destruction since it was created via K_GLOBAL_STATIC, which meant that some of the children within ResourcesModel's QObject tree had already been deleted. This caused a "pure virtual method called" when ResourcesModel was destructed, especially if another thread had been using the QApt::Backend at the time when the QApplication was destroyed.
To fix this, consumers of ResourcesModel now initialize ResourcesModel on their own the first time they use it, parenting it within the QApplication's parentage tree. This ensures that ResourcesModel is destroyed at the appropriate time.

This brings us to the second bug, where the ApplicationBackend didn't lock access to the quit function properly when it was using APT resources, which are thread unsafe, in a separate thread.
Before the first bug was fixed, QApt object deletion was delayed until libc's at_exit exit handler destroyed ResourcesModel & children, which was only called after the initialization thread had finished.

BUG:315003