Make the resources model back a singleton
Centralize the integrateMainWindow in the resources model, so we'll
attach the main window to the model and the model will be in
charge of setting the main window to the backends.
Additionally use this to set the parent of the resources model. No
need to explicitly initialize the model anymore.

CCMAIL: echidnaman@kubuntu.org