Optimize proxy filtering and search
Right now we were calling ResourcesModel::data() for every call when
filtering, now we fetch the resource and we operate on it.
Also when searching fetch QList<AbstractResource*> instead of a list of
package names. This is better because resource pointers are faster to deal
with: no need to see if it matches, comparison is faster too.

This leads to a ~50% optimization while searching (added QTimers to
ResourcesProxyModel::search()).