Reintroduce a busy-wait to releasechecker and document why!
Unfortunately due to incredibly crappy design of MetaReleaseCore we cannot
avoid the pointless context switches to check for .downloading. As
download() is started in a thread and we have no external control over
the thread, it is autostarted in __init__ and there are no synchronization
methods provided by MetaReleaseCore. Calling .download() manually which
would provide a nicely sync function will result in unprotected threading
leading to incorrect upgrade notifications again.

CCBUG: 326172