Change how we manage to get the clean cache
At the moment we were taking advantage of a bug (or anomaly) in QApt where
packages market as ToKeep are not being listed as marked, so this would provide
us proper UI feedback but then when intending to work it wasn't working properly.

This patch instead considers a clean cache the one received on ::prepare()
instead. I was able to reproduce the bug, now I can't reproduce anymore.

BUG: 330244