Make sure we don't provide null packages
Sometimes, apparently, some packages aren't listed when all Application
instances are created. This would become a crash in some special
situations, especially the cases where the architecture of the package
is specified as a suffix to the package name (e.g. primus-libs-ia32).

This patch fixes that crash by not reporting these, but then we still
have the problem that these won't be listed as updates.

CCBUG: 327544