stop installer from exploding when reloading addonswidgets
after an operation was commited we reload the ui which has the unfortunate
side effect that for the briefest of moments we can get a population
request when there is no resource (yet). consequently guard the population
against nullptr access to prevent the installer from crashing.

additional background information: this crash would only occur when one
was looking at an application with addons before the ui reloads (e.g.
if one installs/removes qtcreator from the details page)

BUG: 327998
FIXED-IN: 2.2.1