Port Muon Discover away from PlasmaComponents
Depending on org.kde.plasma.* components is problematic at the moment as
plasma-framework is not coinstallable (or kactivities for that matter) and
it has stopped a couple of people who were trying to contribute to the
project.
Furthermore, it was not a good idea to depend on it anyway. It wouldn't
integrate with the application look and feel settings and we had to juggle
to make it look acceptable.
Counter-side is that most of our components are provided in plasma
framework, so development will be needed to get all needed components in
shape either in Qt or KDeclarative.

REVIEW: 120265
CCMAIL: plasma-devel@kde.org