install releasechecker to libmuon's data dir
releasechecker originally (years ago) was part of the notifier and the
notifier only (back then there was not even a libmuon mind you). since then
the release checking moved into libmuon and libmuonapt, the releasechecker
script however remained in the notifier path.
at this point this is severely misleading for packagers and also just very
wrong.

move the script from the notifier to the libmuon data directory to reflect
that it belongs and is used by the libraries

REVIEW: 122165