Add information about size for not installed apps and runtimes
On top of that I did some code restructuring so some functions
can be easily re-used in case we need to update any part of application.
There is also now runtime-app pairing which is used for calculating
of total size the app will need in case its runtime is not installed.