Download remote icons during initialization of FlatpakResource items
Summary:
Remote icons are used by .flatpakrepo or .flatpak files so we need to download them
to be able to use them later. The icons are downloaded to ~/.cache/discover location
and either used directly if we have downloaded them before or download is initiated in
constructor and once the icon is downloaded we let the UI know to update it.

Reviewers: apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4977