UI feedback when checking for updates
This provides some very light user feedback when the update check
has not been completed. It ensures that we don't get told there
are no updates, when we don't really know yet whether we really do.

End result is that this patch sets the text "Checking for
updates..." on the Update sidebar entry when we're fetching things.

This also introduces a minor fix to the KNS backend, which on
errors would otherwise end up in a perpetual fetching state.

Differential Revision: https://phabricator.kde.org/D5046