Implement UI feedback for long-running searches, and no-results
If a search takes longer than a couple of seconds, a spinner
will face-and-slide in from the bottom of the view. If it then
takes a while longer, a text saying "Still looking" will fade in
very slowly. If the search is near-instant (the general case),
the spinner does not show up.