Queue some connections which cannot be autodetected
Some of the connections we make are doing cross-thread
calls, but in functions which exist on objects which
live inside the same thread. Consequently, connect can't
auto-detect that they need to be queued, and we end up
sometimes (and during searches regularly) crashing
because updates happen at the wrong time. I am unsure
whether this is quite all of them, but with this patch
the search functionality is much less crashy.

Differential Revision: https://phabricator.kde.org/D5409