Use Kirigami.Icon instead of QIconItem
As we are a Kirigami application, this makes use of the Kirigami
icon item, which in turn allows us to use more types of icons,
and not just QIcons. This also allows for remote icons (once a
patch has been added to Kirigami to allow for this, and until
then it fails gracefully by simply failing to load any icon).

Differential Revision: https://phabricator.kde.org/D5769