Simplify KNSBackend fetch logic
This code removes the custom pagination in KNSBackend in favour
of using the internal pagination in KNSCore::Engine. It further
removes the explicit call to request data, as this is already
done by setting the search term, which caused duplicated
results to be returned. Further, remove results already
returned when KNS requests the view to be cleared, further
reducing duplicate view entries.

nb: The fact setting the search term starts a new search is
undocumented, which will need fixing (some thorough
documentation work is ongoing in kns as part of a gsoc project)

BUG: 380138
Differential Revision: https://phabricator.kde.org/D6191