Use appstream-qt insted of glib based one
Summary:
Currently appstream-qt doesn't provide all necessary stuff we need to use and that's why
we used for part of the code glib based appstream API. I implemented the missing bits to
appstream-qt which are now on review and once those changes are merged and new version
of appstream is released we can use appstream-qt only.

Test Plan: I tested the flatpak backend with my changes and it works as before.

Reviewers: apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6267