Search in all app backends in parallel rather than just the preferred
Changes the search strategy into one that will fetch all backends in
parallel. When a same resource is reported twice, the one coming from the
backend selected as preferable for applications will be used.
Duplication matching is done using the appstream id, so it's very important
that application backends support it.