Make sure we load everything in correct order
Summary:
We need to be sure that all apps are reloaded first before checking for updates. This might
be problematic when we are refreshing appstream metadata from remote while simultenously
checking for updates. If updates info arrives earlier then we might end up with no updates
as there are no apps to mark as updateable. We also cannot parse information from flatpak
metadata, this fails as Appstream expect the metadata to have [Desktop entry] group and we
end up with empty component so we need to revert this back to parse desktop file as before.

Reviewers: apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8815