Expose more URLs relevant to the app
Summary:
Make the help, bug tracker, and donation URLs available to the backends, and display them on the App page.

Currently working on a better design to make the metadata section not look so intimidating for apps that behave well and expose all these URLs--but that can go in later, since this has to land on master.

Test Plan:
Tested in KDE Neon with apps available from multiple backends. Here's an example of how it looks with Gedit, which defines all the URLs:

{F5676433}

Reviewers: apol, #discover_software_store

Reviewed By: apol, #discover_software_store

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10131