Improve "show more reviews" string and logic
Summary:
Use slightly more natural text for the "show more reviews" button.

Also fix the bug where the "show more reviews" text would be absent if `rating` was null, which triggered logic that set the number of reviews to be 0. This didn't seem to be adding anything, so I removed it and everything still seems to work fine.

Test Plan:
Browsed a ton of apps. All appeared as expected, including apps like OpenScad that were previously affected by the above bug.

{F5712478}

Reviewers: #discover_software_store, apol

Reviewed By: #discover_software_store, apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10527