Reduce sidebar width
Summary:
BUG: 385992
FIXED-IN: 5.12.3

This is a hack to work around the fact that we're using the wrong UI element for Discover's sidebar (see [[https://bugs.kde.org/show_bug.cgi?id=390928|Bug 390928]]). We use a Kirigami GlobalDrawer for the sidebar, which is problematic because the GlobalDrawer's width is always a minimum of 360 pixels on the desktop--far too wide for our use case (see [[https://bugs.kde.org/show_bug.cgi?id=390927|Bug 390927]]).

Replacing the GlobalDrawer with something more appropriate is going to be a lot of work, so until then, let's not torture our users with the humongous sidebar any longer and just force it to be a smaller size for now.

I think this is a vast visual improvement because it returns the focus to the content, rather than the navigation. It makes Discover feel less "fat."

Test Plan:
Before, small window:
{F5726133}

After, small window:
{F5726132}

Before, medium window:
{F5726113}

After, medium window:
{F5726114}

Before, app page:
{F5726115}

After, app page:
{F5726116}

Before, default view:
{F5726137}

After, default view:
{F5726138}

We may want to consider changing the default window size to be less wide if this patch lands, since with a narrower sidebar, the main view doesn't need to be so wide.

Reviewers: #discover_software_store, apol

Reviewed By: #discover_software_store, apol

Subscribers: acrouthamel, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10756