Improve "Package removal" dialog wording
Summary:
Improve the wording of the "package removal" dialog to feel a bit more natural and introduce less uncertainty.

The list of packages should also be alphabetically sorted, but I couldn't figure out how to do that right now and will do it in a follow-up patch.

Test Plan:
Before:
{F5738085}

After:
{F5738087}

Reviewers: #discover_software_store, apol

Reviewed By: #discover_software_store, apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10992