Show all actions on the main Kirigami toolbar and use the secondary one only for status
Summary:
This is an alternative to D11214 and D11215, and a different method of fixing https://bugs.kde.org/show_bug.cgi?id=391632. I rather prefer it.

The patches moves the {nav Update All} button onto the main Kirigami Toolbar, next to {nav Check for Updates}. This allows the secondary white-background toolbar to be used solely for communicating status information, so there is no longer any possibility of an overlap. It also improves the design IMHO, by putting all the action buttons in one place, and all the status information in another.

BUG: 391632

Test Plan:
All updates selected:
{F5750641}

Some updates not selected:
{F5750640}

Reviewers: apol, #discover_software_store, #vdg

Reviewed By: apol, #discover_software_store

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11218