Unify padding on all sides of app icons in delegates
Summary:
Without this patch, delegate app icons got more padding on the top and bottom than they did on the sides, which looked a little odd and was was especially noticeable with icons that are square or fit into a square.

This patch makes sides of the icon's invisible frame get the same padding, improving the presentation (even for non-square-ish icons, IMHO).

Test Plan:
Normal view:

{F5763961}

{F5763960}

{F5763959}

{F5763958}

Compact view:
{F5763970}

{F5763972}

{F5763971}

{F5763969}

Reviewers: apol, #discover_software_store

Reviewed By: apol, #discover_software_store

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11625