Allow Compact view with narrower windows
Summary:
When the width of the sidebar was reduced, I didn't change the threshold at which the window uses Compact View when it becomes narrower. As a result, Compact View was almost never used in desktop mode, because the threshold at which it triggered was close to the threshold at which the window changes over to mobile view.

This patch corrects the issue and allows the window to switch to Compact View more often.

Test Plan:
Discover's main window can now be made small and narrow and it will switch more quickly to Compact view:

{F5891140}

Looks really really good now IMHO!

Reviewers: apol, #discover_software_store

Reviewed By: apol, #discover_software_store

Subscribers: sharvey, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D13361