Improve the clarity and presentation of the update details text field
Summary: This patche makes it a bit more obvious what's actually going to happen if you click an update to see its details; you get a clear indication of the current version and the new version in the order that your brain expects them to be in.

Test Plan:
Tested in KDE Neon. Before:
{F5891501}

After:
{F5891502}

Needs testing on other platforms that may format things differently.

Reviewers: apol, #discover_software_store

Reviewed By: apol, #discover_software_store

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D13367