Fix Build Fails Due to variable Not set in FINDLIBFWUPD.cmake
Summary: There was no check condition, that the variables in FINDLIBFWUPD.cmake file is set or unset. This diff checks for it. Also, there was a discrepancy for fwupd lib to required or optional, this fixes it. After applying this patch, the build errors are fixed.

Reviewers: apol, davidedmundson

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma, #discover_software_store

Differential Revision: https://phabricator.kde.org/D14566