Allow clicking of links in app descriptions
Prior to this patch, links found in KNewStuff content
descriptions were ignored and both unclickable and
unselectable. With this patch, clicking a link will
launch the default browser pointed at that link.
Currently this is limited to KNS content, but it would
be reasonably trivial to add to other content types as
well, but thought it would be best to discuss on one
type before doing them all at once.

This patch further fixes a minor oversight in the
BBCode neutraliser in the KNS backend (which would
previously eat chunks it was not supposed to, such
as parts of URLs).

BUG: 401176
Differential Revision: https://phabricator.kde.org/D17050