Search-enable the application page
Usually the application page is not the top level page (you'll usually
navigate through it using one of the app lists), but when it is, this
patch ensures that the search field is displayed, and further ensures
that the search is performed on the full list of resources. This will
happen when Discover is launched e.g. from krunner or otherwise
requested to show a specific application on launch.

Differential Revision: https://phabricator.kde.org/D18320