Prettify and search enable the error page
The prettification is based somewhat on the way the Updates page
looks while fetching (it was previously just a header and just
sort of... boring looking)

The functional difference in this patch is that the search field
is enabled, should this page be shown as the first thing when
Discover opens (and as such is the root item in the pagerow). This
might for example happen if you attempt to pass an application on
the commandline which does not exist.

Differential Revision: https://phabricator.kde.org/D18318