UpdatesPage: Fix icon sizing on small screens
Summary:
On Small screens (e.g phones), not all list items have the same height, for example when a package has a long version string.
Since the width of the icon was previously copied from the height, this resulted in inconsistent icon sizes.

This patch sets the icon size to a constant value, which fixes the problem.

BUG: 404070

Test Plan: Icon sizes are consistent and still correctly sized

Reviewers: #discover_software_store

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18964