Add (and use) a LinkLabel component
The component, which is deliberately very simple. just adds the
functionality to Label of ensuring we have a hand cursor when
hovering over a link in the Label component. This is required
due to missing functionality in both Text and QCC2's Label
components, which have no mouse cursor changing abilities. There
is a Qt bug regarding the issue, mentioned in this new
component's code, but so far it looks unlikely to get a "real"
fix anytime soon, so, for less headachiness, this component
(which simply takes code we already had in one place and
applies it slightly more widely and generically)...

BUG:404309
Differential Revision: https://phabricator.kde.org/D19118