[Updates page] Try our best to not elide version strings
Summary:
With this patch, Discover will try its back to never elide version strings on the
Updates page. It will first fall back to a shorter squeezed string that only shows the
available version, then elide from the left so that the most important part doesn't get
cut off.

All of this is mostly only a problem on KDE Neon when using widescreen or mobile view,
but it's nice to fix anyway.

BUG: 404624
FIXED-IN: 5.15.2

Test Plan: {F6628861}

Reviewers: #discover_software_store, colomar, apol

Reviewed By: #discover_software_store, apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19210