Add support for applications in KNSBackend
Prior to this patch, when loading a knsrc file which would
show applications from the KDE Store, the entries would be
shown in the Application Addons menu. This patch adds
functionality to put those into the Application category,
and further adds filtering on architecture for those packages
(so we only show entries with downloadable content supported
by the machine we are running on).

Differential Revision: https://phabricator.kde.org/D19837