Adapt leaf Category names for kns backend, and ensure correct filters
Previously, categories would use the name in KNSRC files (which was
not intended as a display name as much as a handy human friendly
index, in place of the database indices used on the remote server),
and this patch adapts the categories to reflect the intended display
name for those categories after the knewstuff engine has been
initialised. It further ensures that categories are set correctly
for kns resources upon creation.

Differential Revision: https://phabricator.kde.org/D19899