Adapt an x option for app-carrying knsrc files
This adds in an extended option in KNSRC files, which allows them
to tell us whether or not they carry applications. It would also
be possible to make this change in knewstuff itself, but as this
would be somewhat invasive for, frankly, not a great deal of gain,
i propose we do it here. @apol and i discussed this (very) briefly
in person a couple of weeks ago, and it still seems like it would
be reasonable to do it here rather than in the framework.

Differential Revision: https://phabricator.kde.org/D20406