packagekit: Don't display packages that extend themselves
It's common to have appstream components for plugins that some
distributions will split into separate packages. This is useful but it
can easily clutter the UX.
Take Okular for example, it has an appstream component for every format
it supports. On some distributions, like ArchLinux, it will not be
possible to give them separate states as they all point to the same
package. In such case we'll just show the package that is being
extended.

CCBUG: 403118