Add DesktopEntry to notifyrc
This allows the new notification KCM to identify Discover as an application.
While technically the updater isn't "Discover", it is as far as the user is concerned.
If the entry were "System Updates" I would have kept them under Services but since this explicitly says "Discover" which the user should
know is for installing and updating things, moving "Discover" under the application settings makes sense.

Differential Revision: https://phabricator.kde.org/D20919