Use a more conventional sidebar header apparance in desktop view
Summary:
Discover uses the `Kirigami.GlobalDrawer` for its navigation sidebar. This component is
designed to be used for showing global actions and hiding when not in use, and displays
a big banner image on top. This banner concept is borrowed from Android, which has the
same kind of global drawers full of actions with banners on top. So the concept is
somewhat familiar on mobile. However on the desktop, this "sidebar with banner image on
top" UI is not common and feels out of place, giving Discover an odd default appearance
since the sidebar is permanently visible but doesn't look like other apps' sidebars.

This patch adjusts the sidebar appearance to look and feel more conventional in the
desktop view, keeping the banner version only for the mobile view.

Test Plan:
Desktop view: {F6826154}
Mobile view: {F6826086}
(The extra space under the banner image is a pre-existing bug not introduced with this patch)

Now Discover's sidebar looks quite similar to System Settings' sidebar, which is very
visually appealing: {F6826158}

Reviewers: #vdg, #discover_software_store, apol, ndavis

Reviewed By: #vdg, #discover_software_store, apol, ndavis

Subscribers: filipf, ndavis, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21234