Move the notifier logic into a separate process
Summary:
At the moment, the bit that notifies about system updates was a plasmoid and it was working well. The problem was that as a plasmoid it did pull quite a lot of dependencies (flatpak, packagekit) that we kept alive and sometimes were found to de-stabilize the plasmashell process.
This patch moves this into a SNI, the UX stays the same, if it crashes, only this will crash.

Test Plan: Been using it, seems to work okay.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23469