Don't require QtGui for the PaginateModel test
Summary:
This is done by switching from the QStandardItemModel, which requires
QtGui, to QStringListModel, which is in QtCore. The result is a much
leaner test, and it's all done by adding a couple of helper functions
which add the simplistic data addition functions from QSIM, except
locked to QStrings (the way it was being used anyway).

Test Plan: Run the tests, which still pass

Reviewers: #discover_software_store, bshah, apol

Reviewed By: #discover_software_store, apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24156