Revert change to make the notification persistent
Summary:
Nobody liked it and it was annoying. From a conceptual standpoint, it was redundant
anyway because the SNI in the system tray is already persistent when there's an update
available.

Targeting 5.17.1 for this.

Test Plan: Restart plasmashell, see that the "updates available" notification disappears after a few seconds

Reviewers: apol, broulik, #discover_software_store

Reviewed By: apol, #discover_software_store

Subscribers: ognarb, nicolasfella, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24766