Re-add category display to app delegates
Summary:
A year ago, I removed this in 0911fbcd561c with no review. :/ Sorry about that. In the
past year, Discover's UI has come a long way, and I think we can re-add the category
display now without causing the app delegates to become super cluttered.

The category name is suppressed if it would be identical to the page's title.

Test Plan:
Categories in search lists: {F7641610}
Categories on the featured page: {F7641609}
Categories in browse lists, featuring suppression-when-equal-to-the-page-title: {F7641608}

Reviewers: #vdg, #discover_software_store, apol

Reviewed By: #discover_software_store, apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24832